Added support for multiple notification
Only the most recent one will be shown, but not yet dismissed ones will show again after dismissing the current one, since all notifications are now stored on a stack.
Notifications are now saved with a name attribute to facilitate this.
Thus show now takes a name as well.
Accodingly the hide function now takes an argument 'name' to identify which notification to remove from the stack.